+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
+- `set_cache_control_headers` to set `Cache-Control` headers for unpublished records
 - console "CurrentUser" controller allowing users to change their e-mail and password
 
 ### Changed